pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grainger-mro-intelligence"
version = "1.0.0"
description = "MRO Market Intelligence Report Generator for Grainger - integrating ErgoMind, GTA, and FRED data sources with AI-enhanced content generation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
authors = [
    {name = "Grainger Intelligence Team", email = "intelligence@grainger.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3.11",
    "Private :: Do Not Upload",
]

dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "aiohttp==3.9.1",
    "python-docx==1.1.0",
    "backoff==2.2.1",
    "websockets==12.0",
    "python-multipart==0.0.6",
    "python-dateutil==2.8.2",
    "pydantic==2.5.2",
    "anthropic==0.39.0",
    "gunicorn==21.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "types-python-dateutil>=2.8.0",
]

[project.scripts]
mro-intelligence = "mro_intelligence.cli:main"
mro-web = "mro_intelligence.web.app:main"

[project.urls]
Homepage = "https://www.grainger.com"
Documentation = "https://github.com/grainger/mro-intelligence-report-generator"
Repository = "https://github.com/grainger/mro-intelligence-report-generator"

[tool.setuptools.packages.find]
where = ["."]
include = ["mro_intelligence*"]
exclude = ["tests*"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=mro_intelligence",
    "--cov-report=html",
    "--cov-report=term-missing:skip-covered",
    "-v"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["mro_intelligence"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if False:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503", "E501", "E402"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101"]  # Allow assert statements in tests

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["mro_intelligence"]
